requirements.txt
notion-client>=2.2.1
jinja2>=3.1.4
python-dotenv>=1.0.1
requests>=2.32.0
